Added the 'copy-and-enable-user-executable' and 'copy-and-enable-user-executable-result' macros.
These copy a regular file and enable user execute permissions on the destination. These are most commonly used for scripts (for example, bash, perl, python, etc.).

The '-result' variant works against a target dependency ($(<)) as the source and its goal ($(@)) as the destination whereas the "bare" variant works against an explicit source and destination.